Configure storage using conditional compilation
commit-id:8bd780d3
