+++ packages/account-wasm/src/lib.rs
@@ -2,6 +2,5 @@ mod account;
 mod errors;
 mod session;
 mod signer;
-mod storage;
 mod types;
 mod utils;
