+++ packages/account_sdk/Cargo.toml
@@ -38,9 +38,6 @@ thiserror.workspace = true
 toml.workspace = true
 u256-literal.workspace = true
 url.workspace = true
-wasm-bindgen-futures.workspace = true
-wasm-bindgen.workspace = true
-web-sys = "0.3.70"
 indexmap.workspace = true
 num-traits.workspace = true
 tsify-next = "0.5.4"
@@ -63,5 +60,16 @@ serde_json.workspace = true
 tokio.workspace = true
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
+wasm-bindgen-futures.workspace = true
+wasm-bindgen.workspace = true
 wasm-bindgen-test = "0.3.34"
 js-sys = "0.3.70"
+web-sys = { version = "0.3", features = [
+    "CredentialCreationOptions",
+    "CredentialRequestOptions",
+    "CredentialsContainer",
+    "Navigator",
+    "PublicKeyCredential",
+    "PublicKeyCredentialCreationOptions",
+    "Storage",
+] }
