+++ packages/account_sdk/src/tests/owner_test.rs
@@ -6,7 +6,6 @@ use crate::{
     signers::{
         webauthn::WebauthnSigner, HashSigner, NewOwnerSigner, SignError, Signer, SignerTrait,
     },
-    storage::InMemoryBackend,
     tests::{account::webauthn::SoftPasskeySigner, ensure_txn, runners::katana::KatanaRunner},
 };
 use cainome::cairo_serde::{ContractAddress, U256};
@@ -304,7 +303,6 @@ async fn test_change_owner_invalidate_old_sessions() {
         new_signer.clone(),
         controller.address(),
         runner.client().chain_id().await.unwrap(),
-        InMemoryBackend::default(),
     );
 
     let session_account = controller
