+++ packages/account_sdk/src/upgrade.rs
@@ -1,16 +1,13 @@
 use starknet::core::types::Call;
 use starknet_crypto::Felt;
 
-use crate::{controller::Controller, Backend};
+use crate::controller::Controller;
 
 #[cfg(test)]
 #[path = "upgrade_test.rs"]
 mod upgrade_test;
 
-impl<B> Controller<B>
-where
-    B: Backend + Clone,
-{
+impl Controller {
     pub fn upgrade(&self, new_class_hash: Felt) -> Call {
         self.contract().upgrade_getcall(&new_class_hash.into())
     }
